FE : unused files deleted from utils, GraphControl leaned out, functions all exported from utils/index.js
