--- client/components/utils/renderGraph.js
@@ -1,24 +1,0 @@
-// const dictionary = {
-//   bar: <BarGraph data={data} dataset={dataset} x={x} y={y} />,
-//   line: <LineGraph data={data} dataset={dataset} x={x} y={y} />,
-//   scatter: <ScatterChart data={data} dataset={dataset} x={x} y={y} />,
-//   pie: <PieGraph data={data} dataset={dataset} x={x} y={y} />,
-// };
-
-// const renderGraph = (graphTypes) => {};
-
-{
-  /*
-            {graphSelected === 'bar' ? (
-              <BarGraph data={data} dataset={dataset} x={x} y={y} />
-            ) : graphSelected === 'line' ? (
-              <LineGraph data={data} dataset={dataset} x={x} y={y} />
-            ) : graphSelected === 'scatter' ? (
-              <ScatterChart data={data} dataset={dataset} x={x} y={y} />
-            ) : graphSelected === 'pie' ? (
-              <PieGraph data={data} dataset={dataset} x={x} y={y} />
-            ) : (
-              ''
-            )}
-            */
-}
